#400 revert: 정렬 배열의 열 배열 전환을 객체 배열로 되돌린다 — 평점 int 하나는 남기고 이름은 ratingToInt로

+++ src/main/java/org/sopt/solply_server/domain/place/cache/PlaceListSnapshotLoader.java
@@ -411,8 +411,8 @@ public class PlaceListSnapshotLoader {
                 ((Number) row[2]).longValue(),
                 ((Number) row[3]).doubleValue(),
                 toLocalDateTime(row[4]).toEpochSecond(ZoneOffset.UTC),
-                ((Number) row[5]).intValue(),
-                ((Number) row[6]).intValue(),
+                ((Number) row[5]).longValue(),
+                ((Number) row[6]).longValue(),
                 ratingToInt,
                 toNullableDouble(row[8]),
                 toNullableDouble(row[9]));
